rename sm_utils to utils
