+++ packages/chain/statemanager/utils/interface.go
@@ -1,4 +1,4 @@
-package sm_utils
+package utils
 
 import (
 	"github.com/iotaledger/wasp/packages/gpa"
