+++ packages/chain/statemanager/utils/node_randomiser_test.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-package sm_utils
+package utils
 
 import (
 	"testing"
